feat(ai): 新增AI商品顾问功能页面并重构后端服务
- 添加AI商品顾问前端页面(ai-consultant.html)，包含聊天界面和商品推荐模块
- 重命名ConsultantApplication为AiServiceApplication并启用Feign客户端
- 配置应用连接Milvus向量数据库替换Redis，更新API密钥和模型版本
- 修改ChatController支持SSE流式响应，实现TokenStream处理
- 配置Nacos服务发现和MySQL数据源，调整Redis连接参数
- 添加商品推荐API端点，实现基于用户请求的智能商品推荐功能

+++ cart-service/src/main/java/com/hmall/cart/service/impl/CartServiceImpl.java
@@ -155,14 +155,14 @@ public class CartServiceImpl extends ServiceImpl<CartMapper, Cart> implements IC
     }
 
     private void checkCartsFull(Long userId) {
-        int count = lambdaQuery().eq(Cart::getUserId, userId).count();
+        long count = lambdaQuery().eq(Cart::getUserId, userId).count();
         if (count >= cartProperties.getMaxItems()) {
             throw new BizIllegalException(StrUtil.format("用户购物车课程不能超过{}", cartProperties.getMaxItems()));
         }
     }
 
     private boolean checkItemExists(Long itemId, Long userId) {
-        int count = lambdaQuery()
+        long count = lambdaQuery()
                 .eq(Cart::getUserId, userId)
                 .eq(Cart::getItemId, itemId)
                 .count();
